feat(ui-compo): add component based on ant design alert
re #282
