+++ app/renderer/ui/components/alert/index.tsx
@@ -1,24 +1,7 @@
-import * as React from "react"
+import AlertDefault from "./default"
+import AlertMessage from "./message"
 
-export interface Iprops {
-  className?: string
-}
-
-/**
- * Alert UI component
- *
- * @export
- * @class Alert
- * @extends {React.Component<Iprops>}
- */
-
-export default class Alert extends React.Component<Iprops> {
-  // tslint:disable-next-line: completed-docs
-  public render() {
-    return (
-      <div className={this.props.className}>
-        {this.props.children}
-      </div>
-    )
-  }
+export {
+  AlertDefault,
+  AlertMessage
 }
