+++ app/renderer/ui/components/alert/message/style.scss
@@ -1,0 +1,5 @@
+:local {
+  .hidden {
+    display: none;
+  }
+}
